refactor(): 지원서 조회 관련 api 컨트롤러 이동 및 url 변경사항 반영
- 지원자 정보도 함께 주는 경우 `/applicants` 추가
