+++ src/main/kotlin/apply/ui/api/ApplicantRestController.kt
@@ -1,6 +1,5 @@
 package apply.ui.api
 
-import apply.application.ApplicantAndFormResponse
 import apply.application.ApplicantAuthenticationService
 import apply.application.ApplicantResponse
 import apply.application.ApplicantService
@@ -13,7 +12,6 @@ import apply.domain.applicant.Applicant
 import apply.security.LoginApplicant
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -22,32 +20,32 @@ import org.springframework.web.bind.annotation.RestController
 import javax.validation.Valid
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/applicants")
 class ApplicantRestController(
     private val applicantService: ApplicantService,
     private val applicantAuthenticationService: ApplicantAuthenticationService,
     private val mailService: MailService
 ) {
-    @PostMapping("/applicants/register")
+    @PostMapping("/register")
     fun generateToken(@RequestBody @Valid request: RegisterApplicantRequest): ResponseEntity<ApiResponse<String>> {
         val token = applicantAuthenticationService.generateToken(request)
         return ResponseEntity.ok().body(ApiResponse.success(token))
     }
 
-    @PostMapping("/applicants/login")
+    @PostMapping("/login")
     fun generateToken(@RequestBody @Valid request: AuthenticateApplicantRequest): ResponseEntity<ApiResponse<String>> {
         val token = applicantAuthenticationService.generateTokenByLogin(request)
         return ResponseEntity.ok().body(ApiResponse.success(token))
     }
 
-    @PostMapping("/applicants/reset-password")
+    @PostMapping("/reset-password")
     fun resetPassword(@RequestBody @Valid request: ResetPasswordRequest): ResponseEntity<Unit> {
         val newPassword = applicantService.resetPassword(request)
         mailService.sendPasswordResetMail(request, newPassword)
         return ResponseEntity.noContent().build()
     }
 
-    @PostMapping("/applicants/edit-password")
+    @PostMapping("/edit-password")
     fun editPassword(
         @RequestBody @Valid request: EditPasswordRequest,
         @LoginApplicant applicant: Applicant
@@ -56,16 +54,7 @@ class ApplicantRestController(
         return ResponseEntity.noContent().build()
     }
 
-    @GetMapping("/recruitments/{recruitmentId}/application-forms")
-    fun findAllByRecruitmentIdAndKeyword(
-        @PathVariable recruitmentId: Long,
-        @RequestParam applicantKeyword: String?
-    ): ResponseEntity<ApiResponse<List<ApplicantAndFormResponse>>> {
-        val applicants = applicantService.findAllByRecruitmentIdAndSubmittedTrueAndKeyword(recruitmentId, applicantKeyword)
-        return ResponseEntity.ok(ApiResponse.success(applicants))
-    }
-
-    @GetMapping("/applicants")
+    @GetMapping
     fun findAllByKeyword(
         @RequestParam keyword: String
     ): ResponseEntity<ApiResponse<List<ApplicantResponse>>> {
