+++ src/test/kotlin/apply/ui/api/ApplicationFormRestControllerTest.kt
@@ -1,10 +1,12 @@
 package apply.ui.api
 
+import apply.application.ApplicantAndFormResponse
 import apply.application.ApplicantService
 import apply.application.ApplicationFormResponse
 import apply.application.ApplicationFormService
 import apply.application.MyApplicationFormResponse
 import apply.application.mail.MailService
+import apply.createApplicant
 import apply.createApplicationForm
 import apply.createApplicationForms
 import apply.domain.applicant.Applicant
@@ -18,6 +20,7 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.http.HttpHeaders.AUTHORIZATION
+import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.get
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers
 import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder
@@ -44,6 +47,8 @@ internal class ApplicationFormRestControllerTest(
     @MockkBean
     private lateinit var applicantService: ApplicantService
 
+    private val recruitmentId = 1L
+
     private val applicant = Applicant(
         name = "홍길동1",
         email = "applicant_email@email.com",
@@ -60,6 +65,21 @@ internal class ApplicationFormRestControllerTest(
 
     private val myApplicationFormResponses = createApplicationForms().map(::MyApplicationFormResponse)
 
+    private val applicantKeyword = "아마찌"
+
+    private val applicantAndFormResponses = listOf(
+        ApplicantAndFormResponse(
+            createApplicant(name = "로키"), false,
+            createApplicationForms()[0]
+        ),
+        ApplicantAndFormResponse(
+            createApplicant(name = applicantKeyword), false,
+            createApplicationForms()[1]
+        )
+    )
+
+    private val applicantAndFormFindByApplicantKeywordResponses = listOf(applicantAndFormResponses[1])
+
     @BeforeEach
     internal fun setUp(webApplicationContext: WebApplicationContext) {
         mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
@@ -73,10 +93,9 @@ internal class ApplicationFormRestControllerTest(
         every { jwtTokenProvider.isValidToken("valid_token") } returns true
         every { jwtTokenProvider.getSubject("valid_token") } returns applicant.email
         every { applicantService.getByEmail(applicant.email) } returns applicant
-        every { applicationFormService.getApplicationForm(applicant.id, 2L) } returns applicationFormResponse
+        every { applicationFormService.getApplicationForm(applicant.id, recruitmentId) } returns applicationFormResponse
 
-        mockMvc.get("/api/application-forms") {
-            param("recruitmentId", "2")
+        mockMvc.get("/api/recruitments/{recruitmentId}/application-forms", recruitmentId) {
             header(AUTHORIZATION, "Bearer valid_token")
         }.andExpect {
             status { isOk }
@@ -91,11 +110,67 @@ internal class ApplicationFormRestControllerTest(
         every { applicantService.getByEmail(applicant.email) } returns applicant
         every { applicationFormService.getMyApplicationForms(applicant.id) } returns myApplicationFormResponses
 
-        mockMvc.get("/api/application-forms/me") {
+        mockMvc.get("/api/recruitments/{recruitmentId}/application-forms/me", recruitmentId) {
             header(AUTHORIZATION, "Bearer valid_token")
         }.andExpect {
             status { isOk }
             content { json(objectMapper.writeValueAsString(ApiResponse.success(myApplicationFormResponses))) }
+        }
+    }
+
+    @Test
+    fun `특정 모집 id와 지원자에 대한 키워드(이름 or 이메일)로 지원서 정보들을 조회한다`() {
+        val recruitmentId = applicantAndFormResponses[0].applicationForm.recruitmentId
+
+        every {
+            applicantService.findAllByRecruitmentIdAndSubmittedTrueAndKeyword(
+                recruitmentId,
+                applicantKeyword
+            )
+        } returns applicantAndFormFindByApplicantKeywordResponses
+
+        mockMvc.get(
+            "/api/recruitments/{recruitmentId}/application-forms/applicants",
+            recruitmentId
+        ) {
+            contentType = MediaType.APPLICATION_JSON
+            header(AUTHORIZATION, "Bearer valid_token")
+            param("keyword", applicantKeyword)
+        }
+            .andExpect {
+                status { isOk }
+                content {
+                    json(
+                        objectMapper.writeValueAsString(
+                            ApiResponse.success(
+                                applicantAndFormFindByApplicantKeywordResponses
+                            )
+                        )
+                    )
+                }
+            }
+    }
+
+    @Test
+    fun `특정 모집 id에 지원완료한 지원서 정보들을 조회한다`() {
+        val recruitmentId = applicantAndFormResponses[0].applicationForm.recruitmentId
+
+        every {
+            applicantService.findAllByRecruitmentIdAndSubmittedTrueAndKeyword(
+                recruitmentId,
+                null
+            )
+        } returns applicantAndFormResponses
+
+        mockMvc.get(
+            "/api/recruitments/{recruitmentId}/application-forms/applicants",
+            recruitmentId
+        ) {
+            contentType = MediaType.APPLICATION_JSON
+            header(AUTHORIZATION, "Bearer valid_token")
+        }.andExpect {
+            status { isOk }
+            content { json(objectMapper.writeValueAsString(ApiResponse.success(applicantAndFormResponses))) }
         }
     }
 }
